navegando entre as paginas

+++ src/pages/Repositories.js
@@ -1,22 +1,32 @@
 // listagem de repositorios
 import React, { useEffect, useState } from 'react';
 import * as S from '../styles/repositories'; // S de style
+import { useNavigate } from 'react-router-dom';
 
 function Repositories() {
 
+    const navigate = useNavigate();
     const [ repositories, setRepositories ] = useState([]);
 
     // monitora uma variável e dispara uma função qndo o valor da variável mudar
     useEffect(() => {
         let repositoriesName = localStorage.getItem('repositoriesName');
-        repositoriesName = JSON.parse(repositoriesName);
-        setRepositories(repositoriesName);
-        localStorage.clear();
+
+        if(repositoriesName != null) {
+            repositoriesName = JSON.parse(repositoriesName);
+            setRepositories(repositoriesName);
+            localStorage.clear();
+        } else {
+            navigate('/');
+        }
+
+        
     }, []);
 
     return (
         <S.Container>
             <S.Title>Repositórios</S.Title>
+
             <S.List>
                 { repositories.map(repository => {
                     return (
@@ -24,6 +34,8 @@ function Repositories() {
                     )
                 }) }
             </S.List>
+
+            <S.LinkHome to="/">Voltar</S.LinkHome>
         </S.Container>
     )
